+++ packages/apklis_dio_web_api/lib/apklis_dio_web_api.dart
@@ -2,6 +2,7 @@ library apklis_dio_web_api;
 
 import 'package:apklis_web_api/apklis_api_result.dart';
 import 'package:apklis_web_api/apklis_web_api.dart';
+import 'package:apklis_web_api/models/apklis_error_model.dart';
 import 'package:apklis_web_api/models/models.dart';
 import 'package:dio/dio.dart';
 
@@ -20,10 +21,16 @@ class ApklisDioWebApi extends ApklisWebApi {
         final model = ApklisModel.fromJson(json);
         return ApklisApiResult.success(model);
       } else {
-        return ApklisApiResult.failure(response.statusMessage!);
+        return ApklisApiResult.failure(ApklisErrorModel(
+          statusCode: response.statusCode!,
+          statusMessage: response.statusMessage!,
+        ));
       }
     } catch (e) {
-      return ApklisApiResult.failure(e.toString());
+      return ApklisApiResult.failure(ApklisErrorModel(
+        statusCode: -1,
+        statusMessage: e.toString(),
+      ));
     }
   }
 }