+++ packages/apklis_web_api/lib/models/apklis_error_model.dart
@@ -1,0 +1,9 @@
+class ApklisErrorModel {
+  const ApklisErrorModel({
+    required this.statusCode,
+    required this.statusMessage,
+  });
+
+  final int statusCode;
+  final String statusMessage;
+}